Next and Prev month buttons

+++ software/EveryDayCalendar/src/database.cpp
@@ -23,7 +23,7 @@ uint16_t EEPROM_SIZE = sizeof(uint16_t) + (sizeof(uint8_t) * DAYS_IN_YEAR);
 
 // This function returns the number of days in a given month and year
 // It takes the month and year as parameters
-int monthDays(uint8_t month, uint8_t year)
+int monthDays(uint8_t month, uint16_t year)
 {
   // February
   if (month == 2)
@@ -141,7 +141,7 @@ void printDatabase(uint16_t year)
 }
 
 // Return the day of the year for the given year, month and day
-uint GetDayOfTheYear(uint8_t year, uint8_t month, uint8_t day)
+uint GetDayOfTheYear(uint16_t year, uint8_t month, uint8_t day)
 {
   if (month > 12 || day > 31)
   {
@@ -157,7 +157,7 @@ uint GetDayOfTheYear(uint8_t year, uint8_t month, uint8_t day)
   return dayOfTheYear;
 }
 
-void DatabaseSet(uint8_t year, uint8_t month, uint8_t day, uint8_t offset, bool success)
+void DatabaseSet(uint16_t year, uint8_t month, uint8_t day, uint8_t offset, bool success)
 {
   uint dayOfTheYear = GetDayOfTheYear(year, month, day);
 
@@ -193,12 +193,12 @@ uint8_t DatabaseGetOffsetRaw(uint16_t dayOffset) {
   return database[dayOffset];
 }
 
-uint8_t DatabaseGetRaw(uint8_t year, uint8_t month, uint8_t day)
+uint8_t DatabaseGetRaw(uint16_t year, uint8_t month, uint8_t day)
 {
   uint dayOfTheYear = GetDayOfTheYear(year, month, day);
   return database[dayOfTheYear];
 }
-bool DatabaseGet(uint8_t year, uint8_t month, uint8_t day, uint8_t offset)
+bool DatabaseGet(uint16_t year, uint8_t month, uint8_t day, uint8_t offset)
 {
   return bitRead(DatabaseGetRaw(year, month, day), offset);
 }
